+++ KvStore.Core/Application/KeyValue/Commands/PatchKeyValue/PatchKeyValueCommand.cs
@@ -4,6 +4,6 @@ using KvStore.Core.Application.KeyValue.Responses;
 
 namespace KvStore.Core.Application.KeyValue.Commands.PatchKeyValue;
 
-public sealed record PatchKeyValueCommand(string Key, JsonNode? Delta, long? ExpectedVersion)
+public sealed record PatchKeyValueCommand(string Key, JsonNode? Delta, int? ExpectedVersion)
     : ICommand<KeyValueResponse>;
