+++ KvStore.Core/Application/KeyValue/Responses/KeyValueResponse.cs
@@ -3,7 +3,7 @@ using KvStore.Core.Domain.Entities;
 
 namespace KvStore.Core.Application.KeyValue.Responses;
 
-public sealed record KeyValueResponse(string Key, JsonNode? Value, long Version)
+public sealed record KeyValueResponse(string Key, JsonNode? Value, int Version)
 {
     public static KeyValueResponse FromSnapshot(KeyValueResponseSnapshot snapshot)
         => new(snapshot.Key, snapshot.Value?.DeepClone(), snapshot.Version);
